third_party/WebKit/LayoutTests/fast/scrolling/fixed-position-scroll-into-view-expected.txt

Tests that scrolling to a fixed position element does not scroll the viewport.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS scrollY is 0
PASS scrollY is 200
PASS scrollY is 200
PASS scrollY is 0
PASS successfullyParsed is true

TEST COMPLETE
